requirements.txt
pymongo
minio
python-dotenv
